requirements.txt
playwright
python-dotenv
supabase
schedule